+++ packages/core/src/track-reference/track-reference.utils.ts
@@ -1,5 +1,10 @@
 import { Track } from 'livekit-client';
-import { isTrackReference, TrackReferenceOrPlaceholder } from './track-reference.types';
+import { PinState } from '../types';
+import {
+  isTrackReference,
+  isTrackReferencePlaceholder,
+  TrackReferenceOrPlaceholder,
+} from './track-reference.types';
 
 /** Returns a id to identify the `TrackReference` based on participant and source. */
 export function getTrackReferenceId(trackReference: TrackReferenceOrPlaceholder | number): string {
@@ -18,5 +23,45 @@ export function getTrackReferenceSource(trackReference: TrackReferenceOrPlacehol
     return trackReference.publication.source;
   } else {
     return trackReference.source;
+  }
+}
+
+export function isEqualTrackRef(
+  a?: TrackReferenceOrPlaceholder,
+  b?: TrackReferenceOrPlaceholder,
+): boolean {
+  if (isTrackReference(a) && isTrackReference(b)) {
+    return a.publication.trackSid === b.publication.trackSid;
+  } else if (isTrackReferencePlaceholder(a) && isTrackReferencePlaceholder(b)) {
+    return a.participant.identity === b.participant.identity && a.source === b.source;
+  }
+  return false;
+}
+
+/**
+ * Check if the `TrackReference` is pinned.
+ */
+export function isTrackReferencePinned(
+  trackReference: TrackReferenceOrPlaceholder,
+  pinState: PinState | undefined,
+): boolean {
+  if (pinState === undefined) {
+    return false;
+  }
+  if (isTrackReference(trackReference)) {
+    return pinState.some(
+      (pinnedTrackReference) =>
+        pinnedTrackReference.participant.identity === trackReference.participant.identity &&
+        pinnedTrackReference.publication.trackSid === trackReference.publication.trackSid,
+    );
+  } else if (isTrackReferencePlaceholder(trackReference)) {
+    return pinState.some(
+      (pinnedTrackReference) =>
+        pinnedTrackReference.participant.identity === trackReference.participant.identity &&
+        isTrackReferencePlaceholder(pinnedTrackReference) &&
+        pinnedTrackReference.source === trackReference.source,
+    );
+  } else {
+    return false;
   }
 }
